requirements.txt
mysql-connector==2.1.6
mysql-connector-python==8.0.6
